requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.3
django-import-export==3.0.1
django-multiselectfield==0.1.12
django-rest-swagger==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.1
numpy==1.23.4
odfpy==1.4.1
openapi-codec==1.3.2
openpyxl==3.0.10
pandas==1.5.1
Pillow==9.3.0
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
simplejson==3.18.0
six==1.16.0
sqlparse==0.4.3
tablib==3.2.1
uritemplate==4.1.1
urllib3==1.26.12
xlrd==2.0.1
xlwt==1.3.0
